+++ pages/index.tsx
@@ -10,60 +10,73 @@ import Query from '../components/Query';
 import Result from '../components/Result';
 import UserBox from '../components/UserBox';
 
+const isDevelopment = process.env.NODE_ENV === 'development';
+
+const AppContent = () => {
+  return (
+    <Container
+      maxWidth={false}
+      disableGutters={true}
+      sx={{ display: 'flex', flexDirection: 'column', height: '100vh' }}
+    >
+      <Grid container>
+        <Grid item xs={2}>
+          <Box sx={{ m: 2, mt: 1 }}>
+            <ActiveConnection />
+          </Box>
+        </Grid>
+
+        <Grid item xs={9}>
+          <Box sx={{ m: 1 }}>
+            <Message />
+          </Box>
+        </Grid>
+
+        <Grid item xs={1}>
+          {isDevelopment ? <div /> : <UserBox />}
+        </Grid>
+      </Grid>
+
+      <Grid container>
+        <Grid item xs={9}>
+          <Box sx={{ ml: 1, mr: 2 }}>
+            <Input />
+          </Box>
+        </Grid>
+        <Grid item xs={3}>
+          <Box sx={{ m: 2 }} />
+        </Grid>
+      </Grid>
+
+      <Box sx={{ m: 1 }}>
+        <Result />
+      </Box>
+
+      <Grid container sx={{ flexGrow: 1 }}>
+        <Grid item xs={9} sx={{ display: 'flex', flexDirection: 'column' }}>
+          <GraphBox sx={{ flexGrow: 1 }} />
+        </Grid>
+        <Grid item xs={3}>
+          <Box sx={{ ml: 1 }}>
+            <Query />
+          </Box>
+        </Grid>
+      </Grid>
+    </Container>
+  );
+};
+
 const Home: NextPage = () => {
   return (
-    <ClerkProvider>
-      <Container 
-        maxWidth={false} 
-        disableGutters={true} 
-        sx={{ display: 'flex', flexDirection: 'column', height: '100vh' }}
-      >
-        <Grid container>
-          <Grid item xs={2}>
-            <Box sx={{ m: 2, mt: 1 }}>
-              <ActiveConnection />
-            </Box>
-          </Grid>
-
-          <Grid item xs={9}>
-            <Box sx={{ m: 1 }}>
-              <Message />
-            </Box>
-          </Grid>
-
-          <Grid item xs={1}>
-            <UserBox />
-          </Grid>
-        </Grid>
-
-        <Grid container>
-          <Grid item xs={9}>
-            <Box sx={{ ml: 1, mr: 2 }}>
-              <Input />
-            </Box>
-          </Grid>
-          <Grid item xs={3}>
-            <Box sx={{ m: 2 }} />
-          </Grid>
-        </Grid>
-
-        <Box sx={{ m: 1 }}>
-          <Result />
-        </Box>
-
-        <Grid container sx={{ flexGrow: 1 }}>
-          <Grid item xs={9} sx={{ display: 'flex', flexDirection: 'column' }}>
-            <GraphBox sx={{ flexGrow: 1 }} />
-          </Grid>
-          <Grid item xs={3}>
-            <Box sx={{ ml: 1 }}>
-              <Query />
-            </Box>
-          </Grid>
-        </Grid>
-
-      </Container>
-    </ClerkProvider>
+    <>
+      {isDevelopment ? (
+        <AppContent />
+      ) : (
+        <ClerkProvider>
+          <AppContent />
+        </ClerkProvider>
+      )}
+    </>
   );
 };