+++ components/Message.tsx
@@ -2,7 +2,6 @@ import { Typography } from '@mui/material';
 import { observer } from 'mobx-react-lite';
 import { useStores } from '../store/store-container';
 
-
 const Message = observer(() => {
   const { global: store } = useStores();
 
@@ -16,19 +15,11 @@ const Message = observer(() => {
       </Typography>
     );
   }
-  if (store.loaded) {
-    return (
-      <Typography variant="caption" color="gray">
-        {store.message}
-      </Typography>
-    );
-  } else {
-    return (
-      <Typography variant="caption" color="gray">
-        {store.hintsMessage}
-      </Typography>
-    );
-  }
+  return (
+    <Typography variant="caption" color="gray">
+      {store.message}
+    </Typography>
+  );
 });
 
 export default Message;